+++ lib/game/src/lib.rs
@@ -13,6 +13,7 @@ pub mod entity;
 pub mod entity_collection;
 pub mod event;
 pub mod floor_item;
+pub mod item_description_table;
 pub mod item_name_table;
 pub mod item_resource_table;
 pub mod item_slot_count_table;